[#PRX-250 Possible race condition for receiving n-event-rsp]
fix possible null pointer for file deletion

+++ dcm4chee-proxy-service/src/main/java/org/dcm4chee/proxy/forward/ForwardFiles.java
@@ -1559,7 +1559,7 @@ public class ForwardFiles {
         else
             LOG.debug("{}: failed to delete {}", as, infoFile);
         File path = new File(file.getParent());
-        if (path.list().length == 0)
+        if (path!=null && path.list()!=null && path.list().length == 0)
             path.delete();
     }
